Functions that can be reused in the project are grouped in `util.py`
